Fix some merge problems
