+++ Rocket.Chat/Models/Auth/AuthCanBlockMessage.swift
@@ -1,5 +1,5 @@
 //
-//  Auth+CanBlockMessage.swift
+//  AuthCanBlockMessage.swift
 //  Rocket.Chat
 //
 //  Created by Matheus Cardoso on 3/1/18.
