+++ Rocket.Chat/Models/Auth/AuthCanDeleteMessage.swift
@@ -1,13 +1,15 @@
 //
-//  Auth+CanEditMessage.swift
+//  AuthCanDeleteMessage.swift
 //  Rocket.Chat
 //
 //  Created by Matheus Cardoso on 3/1/18.
 //  Copyright © 2018 Rocket.Chat. All rights reserved.
 //
 
+import Foundation
+
 extension Auth {
-    enum CanEditMessageResult {
+    enum CanDeleteMessageResult {
         case allowed
         case timeElapsed
         case differentUser
@@ -16,7 +18,7 @@ extension Auth {
         case unknown
     }
 
-    func canEditMessage(_ message: Message) -> CanEditMessageResult {
+    func canDeleteMessage(_ message: Message) -> CanDeleteMessageResult {
         guard
             let createdAt = message.createdAt,
             let user = user,
@@ -29,20 +31,24 @@ extension Auth {
             return .notActionable
         }
 
-        if user.hasPermission(.editMessage, realm: self.realm) {
+        if user.hasPermission(.forceDeleteMessage, realm: self.realm) {
             return .allowed
         }
 
         func timeElapsed() -> Bool {
-            if settings.messageAllowEditingBlockEditInMinutes < 1 {
+            if settings.messageAllowDeletingBlockDeleteInMinutes < 1 {
                 return false
             }
 
             return Date.serverDate.timeIntervalSince(createdAt)/60 > Double(settings.messageAllowDeletingBlockDeleteInMinutes)
         }
 
+        if user.hasPermission(.deleteMessage, realm: self.realm) {
+            return timeElapsed() ? .timeElapsed : .allowed
+        }
+
         if message.user != user { return .differentUser }
-        if !settings.messageAllowEditing { return .serverBlocked }
+        if !settings.messageAllowDeleting { return .serverBlocked }
 
         if timeElapsed() { return .timeElapsed }
