+++ Rocket.Chat/Models/Auth/AuthUser.swift
